Compatibility with Haml 6
